requirements.txt
Django==3.1
django-admin-sortable2==0.7.7
django-tinymce==3.0.2
Pillow==7.2.0
python-dotenv==0.14.0
python-slugify==4.0.1
requests==2.24.0
